{{cookiecutter.project_name}}/requirements/requirements_local.txt
-r requirements_base.txt

django-seed

{% if cookiecutter.use_django_debug_toolbar == "y" %}
django-debug-toolbar
{% endif %}
